fix: correct migration for apache and nginx packages (#88)
Currently the Apache2 and Nginx .deb packages only properly work for a
fresh install of edumfa. When updating the packages and database
migration steps are required (e.g. privacyidea -> edumfa) the database
currently just get stamped, which breaks edumfa's database migration.

This pull requests fixes that and migration steps are executed
automatically when updating the package or migration to eduMFA.