fix: move edumfa venv install prefix to rules file
When building your own package as described in the docs, you'd get a
wrong path (and broken packages) as the default is /opt/venvs.
Also "-Sa" is not a valid option for dpkg source, error:
dpkg-source: warning: -Sa is not a valid option for Dpkg::Source::Package::V3::Quilt
I assume this is a misspelling of -sa/-sA (see manpage).
Also fix a warning about the rules file not being executable.