set permissions when creating files not after (#894)
This could prevent an attacker racing to read the file before the
permissions are set - especially the DB dump which is unprotected while
the tar command is running (which could take a long time).
In case of a full export with passwords, this could even mean that any
user could read resolver passwords.